Move SetGreetingMessage.cs under /MessageTemplates
